requirements.txt
numpy
matplotlib
tensorflow
tensorflow_hub
jupyterlab
Pillow
ipywidgets
gdown
imutils
pytube
numpy
pytube
opencv-python
scipy

